+++ src/sg/games/tank/entities/Tower.java
@@ -11,9 +11,8 @@ import com.jme3.scene.Spatial;
  */
 public class Tower extends Unit {
 
-    public Tower(long iid, String type, Spatial spatial) {
+    public Tower(int iid, String type, Spatial spatial) {
         super(iid, "Tower", spatial);
-//        speed = 1;
     }
 
     public static class Builder {
@@ -22,8 +21,6 @@ public class Tower extends Unit {
     @Override
     public void update(float tpf) {
         super.update(tpf);
-//        Vector3f localTranslation = this.spatial.getLocalTranslation().clone();
-//        this.spatial.setLocalTranslation(localTranslation.add(0, 0, speed * tpf));
 
     }
 }
